+++ CryptomatteArnold/cryptomatte/cryptomatte_tests.h
@@ -18,7 +18,7 @@ In this test suite, an assertion failure results in an AiMsgError.
 
 */
 
-#define CRYPTO_TEST_FLAG "run_unit_tests"
+#define CRYPTO_TEST_FLAG AtString("run_unit_tests")
 
 ///////////////////////////////////////////////
 //
@@ -336,14 +336,6 @@ inline void run() {
 }
 
 } // namespace HashingTests
-namespace SystemTests {
-inline void critical_section() {
-    if (!g_critsec_active)
-        AiMsgError("Crit section was not initialized in plugin init.");
-}
-
-inline void run() { critical_section(); }
-} // namespace SystemTests
 
 inline void run_all_unit_tests(AtNode* node) {
     if (node && AiNodeLookUpUserParameter(node, CRYPTO_TEST_FLAG) &&
@@ -352,7 +344,6 @@ inline void run_all_unit_tests(AtNode* node) {
         NameParsingTests::run();
         HashingTests::run();
         MaterialNameTests::run();
-        SystemTests::run();
         AiMsgWarning("Cryptomatte unit tests: Complete");
     }
 }