+++ CryptomatteArnold/uigen.py
@@ -5,7 +5,18 @@ import os
 import sys
 import uuid
 from math import pow
-from string import maketrans
+
+# https://stackoverflow.com/questions/436198/what-is-an-alternative-to-execfile-in-python-3
+def _execfile(filepath, globals=None, locals=None):
+    if globals is None:
+        globals = {}
+    globals.update({
+        "__file__": filepath,
+        "__name__": "__main__",
+    })
+    with open(filepath, 'rb') as file:
+        exec(compile(file.read(), filepath, 'exec'), globals, locals)
+
 
 def enum(*sequential, **named):
    enums = dict(zip(sequential, range(len(sequential))), **named)
@@ -34,8 +45,7 @@ class Group(UiElement):
          ident = name.replace(' ', '')
          ident = ident.replace('_', '')
          ident = ident.lower()
-         temptrans = maketrans('','')
-         self.ident = ident.translate(temptrans,'aeiouy')
+         self.ident = "".join(c for c in ident if c not in 'aeiouy')
          
       else:
          self.ident = ident
@@ -56,8 +66,7 @@ class Tab(Group):
          ident = name.replace(' ', '')
          ident = ident.replace('_', '')
          ident = ident.lower()
-         temptrans = maketrans('','')
-         self.ident = ident.translate(temptrans,'aeiouy')
+         self.ident = "".join(c for c in ident if c not in 'aeiouy')
       else:
          self.ident = ident
 
@@ -294,15 +303,15 @@ def DebugPrintElement(el, d):
    indent = ''
    for i in range(d):
       indent += '\t'
-   print '%s %s' % (indent, el)
+   print('%s %s' % (indent, el))
    if isinstance(el, Group) and el.children:
-      print '%s has %d children' % (el, len(el.children))
+      print('%s has %d children' % (el, len(el.children)))
       for e in el.children:
          DebugPrintElement(e, d+1)
       d -= 1
 
 def DebugPrintShaderDef(sd):
-   print '%s' % sd
+   print('%s' % sd)
    DebugPrintElement(sd.root, 0)
 
 def writei(f, s, d=0):
@@ -1196,15 +1205,15 @@ def remapControls(sd):
 # Main. Load the UI file and build UI templates from the returned structure
 if __name__ == '__main__':
    if len(sys.argv) < 7:
-      print 'ERROR: must supply exactly ui source input and mtd, ae, spdl and args outputs'
+      print('ERROR: must supply exactly ui source input and mtd, ae, spdl and args outputs')
       sys.exit(1)
 
    ui = ShaderDef()
    globals_dict = {'ui':ui}
-   execfile(sys.argv[1], globals_dict)
+   _execfile(sys.argv[1], globals_dict)
 
    if not isinstance(ui, ShaderDef):
-      print 'ERROR: ui object is not a ShaderDef. Did you assign something else to it by mistake?'
+      print('ERROR: ui object is not a ShaderDef. Did you assign something else to it by mistake?')
       sys.exit(2)
 
    WriteMTD(ui, sys.argv[2])