+++ pota/src/lentil.cpp
@@ -1,6 +1,6 @@
 #include <ai.h>
 
-AtCritSec l_critsec;
+AtMutex l_critsec;
 bool l_critsec_active;
 inline bool lentil_crit_sec_init();
 inline void lentil_crit_sec_close();